+++ golem-worker-service-base/src/gateway_binding/mod.rs
@@ -11,11 +11,19 @@ pub(crate) use worker_binding::*;
 mod worker_binding_compiled;
 mod worker_binding;
 
-// A gateway binding is more or less the integration to the backend
-// One of the backend is gole worker. A binding can be a static data too.
-// An example of a static binding is cors pre-flight (which is part of `gateway-plugins`)
-// (refer KONG gateway where CORS is a plugin). Some of these bindings is devoid of Rib
-// scripts as they can be static enough which can be `executed` during registration of API definition itself.
+// A gateway binding is integration to the backend. This is similar to AWS's x-amazon-gateway-integration
+// where it holds the details of where to re-route.
+
+// One of the backends (bindings) is golem worker which is the default one.
+// However, there can be other bindings such as file-server, plugin etc.
+// While plugins primarily exist as a collection within other bindings (Example: cors plugin can exist within worker-binding),
+// a plugin can stay standalone as an integration too if serving an incoming request
+// only needs that plugin and nothing else. 
+// Example: Cors-Preflight request.
+//
+// Internal Detail: For static bindings such as `plugins`, any `Rib` script associated with it can be
+// executed during API definition registration itself, and stored as a static value binding. This is important
+// as we pre-compile and pre-compute wherever we can to make serving the original request faster.
 pub enum GatewayBinding {
     Default(WorkerBinding),
     Plugin(Plugin)
