+++ golem-worker-service-base/src/gateway_plugins/http/mod.rs
@@ -1,2 +1,3 @@
 pub use http_plugin::*;
 mod http_plugin;
+mod cors;
